Refactor: Add index.ts to some directories
* Apply changes to all files

* Unify quotation marks
