+++ src/components/common/SVGIconBtn/SVGIconBtn.styles.ts
@@ -1,4 +1,4 @@
-import styled, { css } from "styled-components";
+import styled, { css } from 'styled-components';
 
 export interface IStyledSVGIconBtn {
     color: string;
