+++ src/components/common/index.ts
@@ -1,0 +1,7 @@
+import SVGIconBtn from './SVGIconBtn/SVGIconBtn';
+import UserAvatar from './UserAvatar/UserAvatar';
+
+export {
+    SVGIconBtn,
+    UserAvatar
+}
