+++ src/constants/login.ts
@@ -1,3 +1,3 @@
-import { ProviderNames } from "~/types/auth.type";
+import { ProviderNames } from '~/types/auth.type';
 
 export const socialProviders: ProviderNames[] = ["Google", "Facebook", "Github"];
