+++ src/hooks/stores/useModalStore.ts
@@ -1,4 +1,4 @@
-import { create } from "zustand";
+import { create } from 'zustand';
 import { devtools } from 'zustand/middleware';
 
 interface IState {
