+++ src/index.tsx
@@ -1,9 +1,9 @@
-import React from "react";
+import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from "~/App";
+import App from '~/App';
 
 const container = document.getElementById('root');
-if (!container) throw new Error('Failed to find the root element');
+if (!container) throw new Error("Failed to find the root element");
 const root = createRoot(container);
 root.render(
     <React.StrictMode>
