+++ src/services/index.ts
@@ -1,0 +1,9 @@
+import { authService } from './authService';
+import { firebaseDB } from './database';
+import { quotesAPI } from './quotesApi';
+
+export {
+    authService,
+    firebaseDB,
+    quotesAPI
+}
