+++ src/styles/common/index.ts
@@ -1,0 +1,11 @@
+import * as colors from './colors';
+import * as sizes from './sizes';
+import GlobalStyle from './GlobalStyle';
+import { PrimaryButton } from './PrimaryButton';
+
+export {
+    colors,
+    sizes,
+    GlobalStyle,
+    PrimaryButton
+}
