+++ src/styles/mixins/flexAligns.ts
@@ -1,4 +1,4 @@
-import { css } from "styled-components";
+import { css } from 'styled-components';
 
 export const flexAligns = (justify = 'center', align = 'center') => css`
     display: flex;
