+++ src/styles/mixins/index.ts
@@ -1,0 +1,7 @@
+import { flexAligns } from './flexAligns';
+import { navContentBox } from './navContentBox';
+
+export {
+    flexAligns,
+    navContentBox
+}
