+++ src/utils/index.ts
@@ -1,0 +1,14 @@
+import { capitalizeFirst } from './capitalizeFirst';
+import { debounce } from './debounce';
+import { downloadImage } from './downloadImage';
+import { saveUserData } from './saveUserData';
+import { getStorageData, saveStorageData } from './sessionStorage';
+
+export {
+    capitalizeFirst,
+    debounce,
+    downloadImage,
+    saveUserData,
+    getStorageData,
+    saveStorageData
+}
